backend/requirements.txt
# Django Core
Django==5.1

# Database
psycopg[binary]>=3.2  # PostgreSQL adapter

# Environment & Configuration
python-dotenv==1.0.1
django-environ==0.11.2

# API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1  # JWT authentication
drf-spectacular==0.27.2  # OpenAPI 3.0 documentation
django-filter==24.3  # URL filtering for DRF

# Development
django-extensions==3.2.3  # Extra management commands

# Testing
pytest==8.3.4
pytest-django==4.9.0

# Code Quality
ruff==0.8.2  # Fast linter
